Create Responsive-Portfolio README.md
Hello, 

This work was my introduction to Bootstrap. I had to make three pages and combine the use of Bootstrap in using their templates, with trying to understand the building blocks. It was difficult at first, in the middle and at the end, but enjoyable! It's a whole new world and language though rooted in previous understanding of HTML and CSS. Thanks to my teachers and classmates, and tutor, I've got a good first run-through. 

I got my elements responsive! I got a little ambitious in parts and went on some wild goose chases. I kept it simple with the Contact and I'm glad. I stretched a little on the Portfolio and added some content like music I'm sure I'll switch for webdesign portfolio content in the refractory to come. 

Good times!
https://gcaplan777.github.io/Responsive-Portfolio/index.html